backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
django-filter>=23.5
gunicorn>=21.2.0
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
twilio>=8.0.0
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
anthropic>=0.18.0
requests>=2.28.0
reportlab>=4.0.0